Reorganize deppack functions a little
