+++ lib/deppack/explore.js
@@ -6,7 +6,7 @@ const detective = require('detective');
 
 const mediator = require('../mediator');
 
-const modules = require('./modules');
+const moduleNaming = require('./module-naming');
 const shims = require('./shims');
 const helpers = require('./helpers');
 const xBrowserResolve = require('./resolve');
@@ -41,8 +41,8 @@ const isPackage = helpers.isPackage;
 const isApp = helpers.isApp;
 
 const shouldIncludeDep = path => {
-  const currRoot = modules.getModuleRootName(path);
-  const pkg = modules.packageJson(sysPath.resolve(path));
+  const currRoot = moduleNaming.getModuleRootName(path);
+  const pkg = helpers.packageJson(sysPath.resolve(path));
   const deps = Object.assign({}, pkg.dependencies || {}, pkg.peerDependencies || {});
   return dep => {
     const root = dep.split('/')[0];
@@ -102,11 +102,12 @@ const exploreDeps = fileList => {
   return x => {
     if (!x) return;
     const path = x.path;
-    const source = envify(x.compiled, {NODE_ENV: mediator.isProduction ? 'production' : 'development'});
-    Object.assign(x, {compiled: source});
 
     if (!isJs(fileList, path) || !mediator.npmIsEnabled) return Promise.resolve(x);
     if (path.indexOf('.json') !== -1) return Promise.resolve(x);
+
+    const source = envify(x.compiled, {NODE_ENV: mediator.isProduction ? 'production' : 'development'});
+    Object.assign(x, {compiled: source});
 
     const allDeps = isVendor(path) ? [] : detective(source);
     const usesProcess = isVendor(path) ? false : shims.shouldIncludeProcess(source);
