+++ lib/deppack/helpers.js
@@ -4,6 +4,8 @@ const sysPath = require('path');
 const fs = require('fs');
 const anymatch = require('anymatch');
 const mediator = require('../mediator');
+const deepExtend = require('../helpers').deepExtend;
+const getModuleRootPath = require('./module-naming').getModuleRootPath;
 
 const fsstat = require('../helpers').promisify(fs.stat);
 const isDir = path => fsstat(path).then(stat => stat.isDirectory(), () => Promise.resolve(false));
@@ -27,4 +29,24 @@ const uniq = list => {
   }, {}));
 };
 
-module.exports = {isDir, makeRelative, isRelative, not, globalPseudofile, isVendor, isApp, isPackage, uniq};
+const applyPackageOverrides = pkg => {
+  const pkgOverride = mediator.overrides[pkg.name];
+
+  if (pkgOverride) {
+    pkg = deepExtend(pkg, pkgOverride);
+  }
+
+  return pkg;
+};
+
+const getDepPackageJson = depPath => {
+  const depJson = require(sysPath.join(depPath, 'package.json'));
+  applyPackageOverrides(depJson);
+  return depJson;
+};
+
+const packageJson = filePath => getDepPackageJson(
+  getModuleRootPath(filePath)
+);
+
+module.exports = {isDir, makeRelative, isRelative, not, globalPseudofile, isVendor, isApp, isPackage, uniq, applyPackageOverrides, packageJson};
