+++ lib/deppack/index.js
@@ -6,11 +6,12 @@ const mediator = require('../mediator');
 const modules = require('./modules');
 const explore = require('./explore');
 const helpers = require('./helpers');
+const promisify = require('../helpers').promisify;
 
 const dir = sysPath.dirname(sysPath.dirname(require.resolve('..')));
-const readFile = (path, callback) => fs.readFile(path, {encoding: 'utf8'}, callback);
+const readFile = promisify((path, callback) => fs.readFile(path, {encoding: 'utf8'}, callback));
 
-const wrapInModule = (filePath, opts) => new Promise((resolve, reject) => {
+const wrapInModule = (filePath, opts) => {
   if (opts == null) opts = {};
   if (opts.paths == null) {
     const ref = process.env.NODE_PATH;
@@ -20,11 +21,8 @@ const wrapInModule = (filePath, opts) => new Promise((resolve, reject) => {
 
   filePath = sysPath.resolve(opts.basedir, filePath);
 
-  return readFile(filePath, (err, src) => {
-    if (err) return reject(err);
-    modules.cacheMain(filePath).then(() => resolve(modules.generateModule(filePath, src)));
-  });
-});
+  return readFile(filePath).then(src => modules.wrap(filePath, src));
+};
 
 const brunchRe = /brunch/;
