+++ lib/deppack/modules.js
@@ -1,87 +1,30 @@
 'use strict';
 
 const sysPath = require('path');
-const deepExtend = require('../helpers').deepExtend;
-const mediator = require('../mediator');
 const helpers = require('./helpers');
 const shims = require('./shims');
+const moduleDefinitions = require('./module-definitions');
+const moduleNaming = require('./module-naming');
 
 const not = helpers.not;
-const slashes = string => string.replace(/\\/g, '/');
+
+
+// Naming
+
+const getModuleRootPath = moduleNaming.getModuleRootPath;
+const getModuleFullRootName = moduleNaming.getModuleFullRootName;
+const generateModuleName = moduleNaming.generateModuleName;
+const generateFileBasedModuleName = moduleNaming.generateFileBasedModuleName;
 
 
 // Helpers
 
-const applyPackageOverrides = pkg => {
-  const pkgOverride = mediator.overrides[pkg.name];
-
-  if (pkgOverride) {
-    pkg = deepExtend(pkg, pkgOverride);
-  }
-
-  return pkg;
-};
-
-const getDepPackageJson = depPath => {
-  const depJson = require(sysPath.join(depPath, 'package.json'));
-  applyPackageOverrides(depJson);
-  return depJson;
-};
-
-const packageJson = filePath => getDepPackageJson(
-  getModuleRootPath(filePath)
-);
+const packageJson = helpers.packageJson;
 
 const relativeToRoot = (filePath, relPath) => sysPath.join(
   getModuleRootPath(filePath), relPath
 );
 
-
-
-// Naming
-
-const getModuleRootPath = path => {
-  const split = path.split(sysPath.sep);
-  const index = split.lastIndexOf('node_modules');
-  const scoped = split[index + 1][0] === '@';
-  return split.slice(0, index + 2 + (scoped ? 1 : 0)).join(sysPath.sep);
-};
-
-const getModuleRootName = path => {
-  const split = path.split(sysPath.sep);
-  const index = split.lastIndexOf('node_modules');
-  const scoped = split[index + 1][0] === '@';
-  return split.slice(index + 1, index + 2 + (scoped ? 1 : 0)).join('/');
-};
-
-const getModuleFullRootName = path => {
-  const split = path.split(sysPath.sep);
-  const indexS = split.indexOf('node_modules');
-  const indexE = split.lastIndexOf('node_modules');
-  const scoped = split[indexE + 1][0] === '@';
-  return split.slice(indexS + 1, indexE + 2 + (scoped ? 1 : 0)).join('/');
-};
-
-const generateModuleName = filePath => generateFileBasedModuleName(filePath).replace(/\/([^/]+)\.(json|js)$/, '/$1');
-
-const generateFileBasedModuleName = filePath => {
-  const rootName = getModuleFullRootName(filePath);
-  const rootPath = getModuleRootPath(filePath);
-  const path = filePath.replace(rootPath, '');
-  return slashes(rootName + path);
-};
-
-
-// Definitions
-
-const definition = (name, exp) => {
-  return `require.register("${name}", function(exports, require, module) {
-  module.exports = ${exp};
-});`;
-};
-
-const aliasDef = (target, source) => `require.alias("${source}", "${target}");`;
-const simpleShimDef = (name, obj) => definition(name, JSON.stringify(obj));
 
 // Hanlding of main file and browser overrides
 
@@ -210,7 +153,7 @@ const getHeader = (moduleName, filePath, source) => {
   })();
 });`);
   } else {
-    return definition(moduleName, source);
+    return moduleDefinitions.definition(moduleName, source);
   }
 };
 
@@ -218,38 +161,9 @@ const generateModule = (filePath, source) => {
   return getHeader(generateFileBasedModuleName(filePath), filePath, source);
 };
 
-const makeRequire = (
-`\nvar __makeRelativeRequire = function(require, mappings, pref) {
-  var none = {};
-  var tryReq = function(name, pref) {
-    var val;
-    try {
-      val = require(pref + '/node_modules/' + name);
-      return val;
-    } catch (e) {
-      if (e.toString().indexOf('Cannot find module') === -1) {
-        throw e;
-      }
-
-      if (pref.indexOf('node_modules') !== -1) {
-        var s = pref.split('/');
-        var i = s.lastIndexOf('node_modules');
-        var newPref = s.slice(0, i).join('/');
-        return tryReq(name, newPref);
-      }
-    }
-    return none;
-  };
-  return function(name) {
-    if (name in mappings) name = mappings[name];
-    if (!name) return;
-    if (name[0] !== '.' && pref) {
-      var val = tryReq(name, pref);
-      if (val !== none) return val;
-    }
-    return require(name);
-  }
+const wrap = (filePath, source) => {
+  return cacheMain(filePath).then(() => generateModule(filePath, source));
 };
-`);
 
-module.exports = {aliasDef, simpleShimDef, applyPackageOverrides, generateModule, generateFileBasedModuleName, getModuleRootName, makeRequire, cacheMain, packageJson, resetCache, requiredAliases};
+
+module.exports = {wrap, resetCache, requiredAliases};
