+++ lib/deppack/process.js
@@ -2,6 +2,7 @@
 
 const shims = require('./shims');
 const modules = require('./modules');
+const moduleDefinitions = require('./module-definitions');
 
 const word = /^[_\w]+$/;
 
@@ -22,20 +23,20 @@ const processFiles = (config, root, files, processor) => {
   const usesProcess = shims.usesProcess();
   const _aliases = Object.assign({}, config.npm.aliases || {}, modules.requiredAliases());
 
-  const aliases = Object.keys(_aliases).map(target => modules.aliasDef(target, _aliases[target]));
+  const aliases = Object.keys(_aliases).map(target => moduleDefinitions.aliasDef(target, _aliases[target]));
 
   const used = shims.getUsedShims();
   const shimDefs = used.map(shims.getShimData).map((mapped, index) => {
     if (!mapped) return;
     const shim = used[index];
-    return modules.simpleShimDef(shim, mapped);
+    return moduleDefinitions.simpleShimDef(shim, mapped);
   }).filter(item => item);
 
   const defs = shimDefs.join('\n');
   const proc = usesProcess ? '\nvar process;' : '';
 
   root.add(`\n(function() {
-var global = window;${defs}${proc}${modules.makeRequire}`);
+var global = window;${defs}${proc}${moduleDefinitions.makeRequire}`);
   files.forEach(processor);
   root.add(aliases.join('\n'));
   if (usesProcess) root.add("process = require('process');");
