+++ lib/deppack/resolve.js
@@ -5,7 +5,7 @@ const trueCasePath = require('true-case-path');
 const _glob = require('glob');
 const glob = require('../helpers').promisify(_glob);
 const mediator = require('../mediator');
-const modules = require('./modules');
+const moduleNaming = require('./module-naming');
 const helpers = require('./helpers');
 const shims = require('./shims');
 
@@ -45,7 +45,7 @@ const buildModMap = () => {
 
 const resolveErrorRe = /Cannot find module '(.+)' from '(.+)'/;
 const friendlyRequireError = (mod, opts, err) => {
-  const topLevel = modules.getModuleRootName(mod);
+  const topLevel = moduleNaming.getModuleRootName(mod);
   if (resolveErrorRe.test(err.message)) {
     const data = err.message.match(resolveErrorRe);
     const mod = data[1];
@@ -80,7 +80,7 @@ const checkImproperCase = (mod, opts, res) => {
 };
 
 const resolve = (mod, opts, cb) => {
-  Object.assign(opts, {packageFilter: modules.applyPackageOverrides, modules: modMap, extensions: ['.js', '.json']});
+  Object.assign(opts, {packageFilter: helpers.applyPackageOverrides, modules: modMap, extensions: ['.js', '.json']});
   mod = mediator.npm.aliases && mediator.npm.aliases[mod] || mod;
   browserResolve(mod, opts, (err, res) => {
     if (err) {
